requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
Django==5.0.4
django-cookies-samesite==0.9.0
django-cors-headers==4.3.1
django-sslserver==0.22
djangorestframework==3.15.1
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
jsonfield==3.1.0
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
phonenumbers==8.13.37
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
pyparsing==3.1.2
requests==2.31.0
rsa==4.9
sqlparse==0.5.0
twilio==9.0.5
typing_extensions==4.11.0
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
